requirements.txt
opencv-python
mss
numpy
pyautogui
pillow
mcp[cli]